latest fixes: toast replaced with on registration alert, button visibility fix

+++ ZbsProjectCMT/CMT Smart Claim/src/spire/example/cmt/Call_me_back.java
@@ -290,7 +290,7 @@ public class Call_me_back extends Activity implements OnClickListener {
 					obj.put("PersonalDataId", new Integer(0));
 
 					Map in_obj = new LinkedHashMap();
-					in_obj.put("Id", new Integer(1));
+					in_obj.put("Id", new Integer(0));
 					in_obj.put("Title", null);
 					in_obj.put("FirstName", names_info[1]);
 					in_obj.put("Surname", names_info[2]);
